+++ src/dendroscope/commands/select/SelectShortEdgesCommand.java
@@ -74,23 +74,20 @@ public class SelectShortEdgesCommand extends CommandBase implements ICommand {
     /**
      * parses the given command and executes it
      *
-     * @param np
-     * @throws java.io.IOException
-     */
+	 */
     public void apply(NexusStreamParser np) throws Exception {
     }
 
     /**
      * action to be performed
      *
-     * @param ev
-     */
+	 */
     public void actionPerformed(ActionEvent ev) {
         float threshold = (float) ProgramProperties.get("ShortEdgeThreshold", 0.1);
         String result = JOptionPane.showInputDialog(getViewer().getFrame(), "Enter maximum edge length to select:", "" + threshold);
         if (result != null && NumberUtils.isFloat(result)) {
-            ProgramProperties.put("ShortEdgeThreshold", (double) threshold);
-            executeImmediately("select edges=short threshold=" + result + ";");
+			ProgramProperties.put("ShortEdgeThreshold", threshold);
+			executeImmediately("select edges=short threshold=" + result + ";");
         }
     }
